Move MultiGenerator to the util package
