+++ src/eradiate/util/misc.py
@@ -580,3 +580,35 @@ def find_runs(
         run_lengths = np.diff(np.append(run_starts, n))
 
         return run_values, run_starts, run_lengths
+
+
+class MultiGenerator:
+    """
+    This generator aggregates several generators and makes sure that items that
+    have already been served are not repeated.
+    """
+
+    def __init__(self, generators):
+        self.generators = generators
+        self._i_generator = 0
+        self._current_iterator = iter(self.generators[self._i_generator])
+        self._visited = set()
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        try:
+            result = next(self._current_iterator)
+            if result not in self._visited:
+                self._visited.add(result)
+                return result
+            else:
+                return self.__next__()
+        except StopIteration:
+            if self._i_generator >= len(self.generators) - 1:
+                raise
+            else:
+                self._i_generator += 1
+                self._current_iterator = iter(self.generators[self._i_generator])
+                return self.__next__()
